moved change goal status
