+++ docs/openapi/paths/index.yaml
@@ -52,8 +52,8 @@
   $ref: './recipient/recipient-search.yaml'
 '/recipient/goals/{recipientId}':
   $ref: './recipient/recipient-goals.yaml'
-'/recipient/goals/{goalId}/changeStatus':
-  $ref: './recipient/recipient-goals-status.yaml'
+'/goals/{goalId}/changeStatus':
+  $ref: './goals/goal-status.yaml'
 '/widgets/overview':
   $ref: './widgets/overview.yaml'
 '/widgets/reasonList':
