+++ src/routes/goals/handlers.js
@@ -1,0 +1,26 @@
+import { updateGoalStatusById } from '../../services/goals';
+import handleErrors from '../../lib/apiErrorHandler';
+
+const namespace = 'SERVICE:GOALS';
+
+const logContext = {
+  namespace,
+};
+
+// eslint-disable-next-line import/prefer-default-export
+export async function changeGoalStatus(req, res) {
+  try {
+    const { goalId } = req.params;
+    const { newStatus } = req.body;
+    const updatedGoal = await updateGoalStatusById(goalId, newStatus);
+
+    if (!updatedGoal) {
+      res.sendStatus(404);
+      return;
+    }
+
+    res.json(updatedGoal);
+  } catch (error) {
+    await handleErrors(req, res, error, logContext);
+  }
+}
