+++ src/routes/goals/index.js
@@ -1,0 +1,6 @@
+import express from 'express';
+import { changeGoalStatus } from './handlers';
+
+const router = express.Router();
+router.put('/:goalId/changeStatus', changeGoalStatus);
+export default router;
