+++ src/routes/recipient/handlers.js
@@ -1,5 +1,5 @@
 import {
-  getGoalsByActivityRecipient, recipientById, recipientsByName, updateRecipientGoalStatusById,
+  getGoalsByActivityRecipient, recipientById, recipientsByName,
 } from '../../services/recipient';
 import handleErrors from '../../lib/apiErrorHandler';
 import filtersToScopes from '../../scopes';
@@ -58,23 +58,6 @@ export async function getGoalsByRecipient(req, res) {
     }
 
     res.json(recipient);
-  } catch (error) {
-    await handleErrors(req, res, error, logContext);
-  }
-}
-
-export async function changeRecipientGoalStatus(req, res) {
-  try {
-    const { goalId } = req.params;
-    const { newStatus } = req.body;
-    const updatedGoal = await updateRecipientGoalStatusById(goalId, newStatus);
-
-    if (!updatedGoal) {
-      res.sendStatus(404);
-      return;
-    }
-
-    res.json(updatedGoal);
   } catch (error) {
     await handleErrors(req, res, error, logContext);
   }
