+++ src/routes/recipient/index.js
@@ -3,12 +3,10 @@ import {
   getRecipient,
   searchRecipients,
   getGoalsByRecipient,
-  changeRecipientGoalStatus,
 } from './handlers';
 
 const router = express.Router();
 router.get('/search', searchRecipients);
 router.get('/:recipientId', getRecipient);
 router.get('/goals/:recipientId', getGoalsByRecipient);
-router.put('/goals/:goalId/changeStatus', changeRecipientGoalStatus);
 export default router;
