+++ src/services/goals.test.js
@@ -1,14 +1,14 @@
 import { Op } from 'sequelize';
 import {
-  copyGoalsToGrants, saveGoalsForReport, goalsForGrants,
+  copyGoalsToGrants, saveGoalsForReport, goalsForGrants, updateGoalStatusById,
 } from './goals';
-
 import {
   Goal,
   Grant,
   Objective,
   ActivityReportObjective,
   GrantGoal,
+  sequelize,
 } from '../models';
 
 describe('Goals DB service', () => {
@@ -230,5 +230,35 @@ describe('goalsForGrants', () => {
       ],
       order: ['createdAt'],
     });
+  });
+});
+
+describe('Change Goal Status', () => {
+  let goal;
+
+  beforeAll(async () => {
+    // Create Goal.
+    goal = await Goal.create({
+      name: 'Goal with Objectives',
+      status: 'Not Started',
+      timeframe: '12 months',
+      isFromSmartsheetTtaPlan: false,
+      createdAt: new Date('2021-01-02'),
+    });
+  });
+
+  afterAll(async () => {
+    // Cleanup Goal.
+    await Goal.destroy({
+      where: {
+        id: goal.id,
+      },
+    });
+    await sequelize.close();
+  });
+  it('Updates goal status', async () => {
+    const newStatus = 'In Progress';
+    const updatedGoal = await updateGoalStatusById(goal.id.toString(), newStatus);
+    expect(updatedGoal.status).toEqual(newStatus);
   });
 });
